refactor: add service to file downloader name
